requirements.txt
absl-py==1.0.0
attrs==21.4.0
autopep8==1.6.0
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
mediapipe==0.8.10
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
playsound==1.2.2
protobuf==3.20.1
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2